FEAT(CanHandlerRunner): Adding the runner
On the present commit the handler runner of all the concrete components
runners is added. If you run the .py file all the components handler
must initialize and run correctly.

The behavior of the handlers whas only hand tested, a complete test on a
real enviroment is necesary to ensure full security.